+++ packages/airnode-deployer/test/fixtures/config/config.ts
@@ -24,6 +24,7 @@ export function buildConfig(overrides?: Partial<Config>): Config {
   return {
     chains: [
       {
+        maxConcurrentTransactions: 100,
         authorizers: [],
         contracts: {
           AirnodeRrp: '0x197F3826040dF832481f835652c290aC7c41f073',
