+++ packages/airnode-examples/integrations/authenticated-coinmarketcap/config.example.json
@@ -1,6 +1,7 @@
 {
   "chains": [
     {
+      "maxConcurrentTransactions": 100,
       "authorizers": [],
       "contracts": {
         "AirnodeRrp": "0x5FbDB2315678afecb367f032d93F642f64180aa3"
