+++ packages/airnode-node/src/evm/handlers/process-transactions.test.ts
@@ -39,7 +39,7 @@ const createConfig = (txType: 'legacy' | 'eip1559') => {
 
 describe('processTransactions', () => {
   test.each(['legacy', 'eip1559'] as const)(
-    'fetches the gas price, assigns nonces and submits transactions - txType: %d',
+    'fetches the gas price, assigns nonces and submits transactions - txType: %s',
     async (txType) => {
       const config = createConfig(txType);
       const { gasTarget, blockSpy, gasPriceSpy } = createAndMockGasTarget(txType);
@@ -171,7 +171,7 @@ describe('processTransactions', () => {
   );
 
   test.each(['legacy', 'eip1559'] as const)(
-    `does not submit transactions if a gas price cannot be fetched - txType: %d`,
+    `does not submit transactions if a gas price cannot be fetched - txType: %s`,
     async (txType) => {
       const config = createConfig(txType);
       const contract = new ethers.Contract('address', ['ABI']);
