+++ packages/airnode-node/src/requests/grouping.ts
@@ -1,6 +1,7 @@
 import groupBy from 'lodash/groupBy';
 import uniq from 'lodash/uniq';
-import { GroupedRequests } from '../types';
+import { AnyRequest, ApiCall, GroupedRequests, Request, RequestType, Withdrawal } from '../types';
+import { removeKey } from '../utils/object-utils';
 
 export interface RequestsBySponsorAddress {
   readonly [sponsorAddress: string]: GroupedRequests;
@@ -31,4 +32,31 @@ export function groupRequestsBySponsorAddress(requests: GroupedRequests): Reques
   }, {} as RequestsBySponsorAddress);
 
   return groupedRequests;
+}
+
+export function flattenRequests(groupedRequests: GroupedRequests): Request<AnyRequest>[] {
+  // Store the type as well temporarily so that requests can be ungrouped again
+  const apiCalls = groupedRequests.apiCalls.map((apiCall) => ({ ...apiCall, __type: RequestType.ApiCall }));
+
+  const withdrawals = groupedRequests.withdrawals.map((withdrawal) => ({
+    ...withdrawal,
+    __type: RequestType.Withdrawal,
+  }));
+
+  // Requests are processed with the following priority:
+  //   1. API calls
+  //   2. Withdrawals
+  return [...apiCalls, ...withdrawals];
+}
+
+export function groupRequests(flatRequests: Request<any>[]): GroupedRequests {
+  const apiCalls = flatRequests
+    .filter((request) => request.__type === RequestType.ApiCall)
+    .map((request) => removeKey(request, '__type')) as Request<ApiCall>[];
+
+  const withdrawals = flatRequests
+    .filter((request) => request.__type === RequestType.Withdrawal)
+    .map((request) => removeKey(request, '__type')) as Request<Withdrawal>[];
+
+  return { apiCalls, withdrawals };
 }
